# Add auto mtls to benchmark tests and improve memory