```
name                       old time/op        new time/op        delta
ClusterGeneration/empty-8        4.84ms ± 0%        4.28ms ± 0%   ~     (p=1.000 n=1+1)

name                       old alloc/op       new alloc/op       delta
ClusterGeneration/empty-8        1.93MB ± 0%        1.71MB ± 0%   ~     (p=1.000 n=1+1)

name                       old allocs/op      new allocs/op      delta
ClusterGeneration/empty-8         23.4k ± 0%         21.0k ± 0%   ~     (p=1.000 n=1+1)
```